requirements.txt
aiohttp==3.6.3
asgiref==3.3.1
async-timeout==3.0.1
attrs==20.3.0
certifi==2020.12.5
chardet==3.0.4
discord==1.0.1
discord.py==1.5.1
Django==3.1.5
idna==2.10
image==1.5.33
multidict==4.7.6
Pillow==8.1.0
pyserial==3.4
pytz==2020.5
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.2
yarl==1.5.1
